Reorganized files. Fixed `UnhandledPromiseRejectionWarning`.
