+++ test/integration/templates/CONSTANTS.js
@@ -1,0 +1,2 @@
+module.exports.DEFAULT_PATH_TO_EXTPECTED_TEMPLATES = './test/integration/templates/expected';
+module.exports.DEFAULT_PATH_TO_ACTUAL_TEMPLATES = './test/integration/templates/actual';
