+++ test/integration/templates/expected/html/no-params/index.html
@@ -692,7 +692,7 @@ integer
     
     <pre class="hljs mb-4 border border-grey-darkest rounded"><code>{
   "lumens": 0,
-  "sentAt": "2019-12-11T14:57:56Z"
+  "sentAt": "2019-12-12T14:53:17Z"
 }</code></pre>
     <h6 class="text-xs font-bold text-grey-darker italic">This example has been generated automatically.</h6>
     
@@ -1222,7 +1222,7 @@ integer
     
     <pre class="hljs mb-4 border border-grey-darkest rounded"><code>{
   "command": "on",
-  "sentAt": "2019-12-11T14:57:56Z"
+  "sentAt": "2019-12-12T14:53:17Z"
 }</code></pre>
     <h6 class="text-xs font-bold text-grey-darker italic">This example has been generated automatically.</h6>
     
@@ -1749,7 +1749,7 @@ integer
     
     <pre class="hljs mb-4 border border-grey-darkest rounded"><code>{
   "command": "on",
-  "sentAt": "2019-12-11T14:57:56Z"
+  "sentAt": "2019-12-12T14:53:17Z"
 }</code></pre>
     <h6 class="text-xs font-bold text-grey-darker italic">This example has been generated automatically.</h6>
     
@@ -2242,7 +2242,7 @@ integer
     
     <pre class="hljs mb-4 border border-grey-darkest rounded"><code>{
   "percentage": 0,
-  "sentAt": "2019-12-11T14:57:56Z"
+  "sentAt": "2019-12-12T14:53:17Z"
 }</code></pre>
     <h6 class="text-xs font-bold text-grey-darker italic">This example has been generated automatically.</h6>
