+++ test/integration/templates/generation scripts/templates/generateHtml.js
@@ -1,5 +1,5 @@
 
-const { generateTemplate } = require('./utils');
+const { generateTemplate } = require('../../../../utils/utils');
 const asyncFile = './test/docs/streetlights.yml';
 
 function generate(pathToTemplates) {
