+++ test/integration/templates/java-spring.spec.js
@@ -1,9 +1,10 @@
 /* eslint-disable no-undef */
 const dircompare = require('dir-compare');
 const options = { compareSize: true };
-const pathToExpectedTemplates = './test/integration/templates/expected/java-spring';
-const pathToactualTemplates = './test/integration/templates/actual/java-spring';
-const generateJavaSpring = require('./generation scripts/utils/generateJavaSpring');
+const constants = require('./CONSTANTS');
+const pathToExpectedTemplates = `${constants.DEFAULT_PATH_TO_EXTPECTED_TEMPLATES}/java-spring`;
+const pathToactualTemplates = `${constants.DEFAULT_PATH_TO_ACTUAL_TEMPLATES}/java-spring`;
+const generateJavaSpring = require('./generation scripts/templates/generateJavaSpring');
 const expect = require('chai').expect;
 const { deleteFolderRecursive } = require('../../utils/utils');
 describe('java-spring', () => {
